+++ artifacttemplates/http%3A%2F%2Fopentosca.org%2Fartifacttemplates/DockerEngine_DockerEngine-Interface-w1/source/src/main/java/org/opentosca/artifacttemplates/Constants.java
@@ -1,4 +1,4 @@
-package com.example.producingwebservice;
+package org.opentosca.artifacttemplates;
 
 public class Constants {
