+++ artifacttemplates/http%3A%2F%2Fopentosca.org%2Fartifacttemplates/DockerEngine_DockerEngine-Interface-w1/source/src/main/java/org/opentosca/artifacttemplates/DockerEngineInterfaceDockerEngineApplication.java
@@ -1,4 +1,4 @@
-package com.example.producingwebservice;
+package org.opentosca.artifacttemplates;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
