+++ artifacttemplates/http%3A%2F%2Fopentosca.org%2Fartifacttemplates/DockerEngine_DockerEngine-Interface-w1/source/src/test/java/org/opentosca/artifacttemplates/DockerEngineInterfaceDockerEngineApplicationTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.producingwebservice;
+package org.opentosca.artifacttemplates;
 
 import static org.assertj.core.api.Assertions.assertThat;
